Create UserManual.md
This is the user's manual for the Color Palette GUI suite that I'm composing for the class SDEV 140 at IvyTech South Bend/Elkhart for Summer of 2025. Like the project itself, the documentation here is a work in progress and will be updated to reflect any changes I make to my file and its content in the future.